requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.4.0
certifi==2022.5.18.1
chardet==4.0.0
charset-normalizer==2.0.12
discord-components==2.1.2
discord.py==1.7.3
idna==3.3
multidict==6.0.2
python-dateutil==2.8.2
requests==2.28.0
six==1.16.0
twitchAPI==2.5.5
typing_extensions==4.2.0
urllib3==1.26.9
websockets==10.3
yarl==1.7.2
